backend/requirements/base.txt
feedparser==6.0.11
psycopg2-binary==2.9.9
sgmllib3k==1.0.0
true-noorm==0.1.2
